refactor(oracle): test names and files
