+++ x/oracle/keeper/params_test.go
@@ -1,0 +1,50 @@
+package keeper
+
+import (
+	"testing"
+
+	"github.com/NibiruChain/nibiru/x/common/asset"
+	"github.com/NibiruChain/nibiru/x/common/denoms"
+	"github.com/NibiruChain/nibiru/x/oracle/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/stretchr/testify/require"
+)
+
+func TestParams(t *testing.T) {
+	input := CreateTestFixture(t)
+
+	// Test default params setting
+	input.OracleKeeper.SetParams(input.Ctx, types.DefaultParams())
+	params := input.OracleKeeper.GetParams(input.Ctx)
+	require.NotNil(t, params)
+
+	// Test custom params setting
+	votePeriod := uint64(10)
+	voteThreshold := sdk.NewDecWithPrec(33, 2)
+	minVoters := uint64(4)
+	oracleRewardBand := sdk.NewDecWithPrec(1, 2)
+	slashFraction := sdk.NewDecWithPrec(1, 2)
+	slashWindow := uint64(1000)
+	minValidPerWindow := sdk.NewDecWithPrec(1, 4)
+	whitelist := []asset.Pair{
+		asset.Registry.Pair(denoms.BTC, denoms.NUSD),
+		asset.Registry.Pair(denoms.ETH, denoms.NUSD),
+	}
+
+	// Should really test validateParams, but skipping because obvious
+	newParams := types.Params{
+		VotePeriod:        votePeriod,
+		VoteThreshold:     voteThreshold,
+		MinVoters:         minVoters,
+		RewardBand:        oracleRewardBand,
+		Whitelist:         whitelist,
+		SlashFraction:     slashFraction,
+		SlashWindow:       slashWindow,
+		MinValidPerWindow: minValidPerWindow,
+	}
+	input.OracleKeeper.SetParams(input.Ctx, newParams)
+
+	storedParams := input.OracleKeeper.GetParams(input.Ctx)
+	require.NotNil(t, storedParams)
+	require.Equal(t, storedParams, newParams)
+}
